+++ Changes.md
@@ -1,7 +1,10 @@
 1.4.x.x (relative to 1.4.9.0)
 =======
 
+Improvements
+------------
 
+- LightEditor : Values of inherited attributes are now displayed in the Light Editor. These are presented as dimmed "fallback" values.
 
 1.4.9.0 (relative to 1.4.8.0)
 =======